+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## [Unreleased][unreleased]
 
 ## Removed
-
 - `createTempUser()`
+
+### Changed
+- new webxdc api signature for `Context.getWebxdcStatusUpdates`
+- Upgrade to deltachat-core version `1.77.0`
 
 ## [1.76.0] - 2022-03-04